+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Set up Algolia credentials;
+
 ## [1.0.2] - 2025-09-25
 ### Added
 - Algolia domain verification;